针对精度转换时的警告进行处理

+++ Driver/chassis.h
@@ -7,9 +7,9 @@
 typedef struct chassis_structure
 {
     bool _switch;//是否使能底盘
-    float x_speed;//x方向底盘速度
-    float y_speed;//y方向底盘速度
-    float w_speed;//w方向底盘角速度
+    double x_speed;//x方向底盘速度
+    double  y_speed;//y方向底盘速度
+    double  w_speed;//w方向底盘角速度
 } CHASSIS;
